Widen the global-cell wire to uint64 (L9+ readiness, plan appendix item 1)
Post-7C particles carry LOCAL cells (bounded per-rank), so only the
GLOBAL encodings sat on the 2^32 wall: wire_cell() now returns uint64,
the migration Alltoallv cell buffers/type are uint64, arrival unpack
decodes at 64 bits, and the checkpoint-restore narrowing guard is gone
(disk and wire are the same width; the on-disk format is unchanged).
The lmesh build aborts loudly if a rank's LOCAL cell space would reach
the uint32 empty_cell sentinel; max_cell() widened to size_t.

Unit test pins the wire encoding above 2^32 with synthetic L10 numbers.
Acceptance baselines unchanged (LIVE 630 / stress 529, MISOWNED 0);
host checkpoint continuity remains BITWISE.  Also recorded appendix
item 1b (int global cochain/vertex indices — the same ~L9 wall in the
layout/plan/geometry code, discovered by this audit, still open) and
raised the checkpoint script's GPU delta to sit above the measured
run-to-run atomic-reorder envelope (~2e-4 in J_e).

+++ src/systems/prismatic/prismatic_ptc_mesh_local.h
@@ -150,10 +150,15 @@ struct prismatic_ptc_mesh_ptrs {
   }
 
   // GLOBAL (rank-agnostic) wire encoding of a LOCAL cell, plan F7.
-  HD_INLINE uint32_t wire_cell(uint32_t cell) const {
+  // 64-bit: the global cell space k·N_tri_global + tri exceeds 2^32
+  // near L9 (checkpoint plan appendix item 1) — the wire and the
+  // checkpoint format carry uint64.  LOCAL cells stay uint32 (bounded
+  // by the per-rank mesh size; guarded loudly at build).
+  HD_INLINE uint64_t wire_cell(uint32_t cell) const {
     const int lay = int(cell) / N_tri;
     const int tri = int(cell) - lay * N_tri;
-    return uint32_t((k0 + lay) * N_tri_global + tri_l2g[tri]);
+    return uint64_t(k0 + lay) * uint64_t(N_tri_global) +
+           uint64_t(tri_l2g[tri]);
   }
 
   // =======================================================================
@@ -292,7 +297,7 @@ class prismatic_ptc_mesh_local {
   int lay_own_lo() const { return m_lay_own_lo; }
   int lay_own_hi() const { return m_lay_own_hi; }
   int n_tri_global() const { return m_n_tri_global; }
-  int max_cell() const { return m_n_tri_local * m_n_layers; }
+  size_t max_cell() const { return size_t(m_n_tri_local) * m_n_layers; }
 
   // World-rank migration parameters (canonical A·K decomposition).
   int n_angular_ranks() const { return m_A; }
